+++ src/pages/DashboardPage/index.js
@@ -1,5 +1,5 @@
 export { default as DashboardPage } from './DashboardPage';
 export { default as DashboardInfoPage } from './DashboardInfoPage/DashboardInfoPage';
-export { default as DashboardMessagePage } from './DashboardMessagePage/DashboardMessagePage';
+export { default as DashboardEmailPage } from './DashboardEmailPage/DashboardEmailPage';
 export { default as DashboardAttendeePage } from './DashboardAttendeePage/DashboardAttendeePage';
 export { default as DashboardStatisticPage } from './DashboardStatisticPage/DashboardStatisticPage';
